Lexer mostly complete. Emit almost complete

+++ EASM/lib/registers.h
@@ -1,9 +1,8 @@
 #ifndef _REGISTERS_H_
 #define _REGISTERS_H_
 
-#include <stdint.h>
 #include <string.h>
 
-int16_t getregister (char * c);
+int getregister (char * c);
 
 #endif
